+++ CHANGELOG.md
@@ -27,6 +27,7 @@ Not yet released.
 * [CS-442][], [CS-443][], [CS-449][] Multiple failures when running XNAT in a non-root context: pulling images, project settings, bulk launching
 * [CS-448][] Fix height of 'Edit Command' dialog so full code editor is displayed
 * [CS-450][] Restrict usage of command automation to commands that match the context of selected events
+* [CS-454][] Container working directory was not being saved at launch
 
 [CS-355]: https://issues.xnat.org/browse/CS-355
 [CS-420]: https://issues.xnat.org/browse/CS-420
@@ -40,6 +41,7 @@ Not yet released.
 [CS-448]: https://issues.xnat.org/browse/CS-448
 [CS-449]: https://issues.xnat.org/browse/CS-449
 [CS-450]: https://issues.xnat.org/browse/CS-450
+[CS-454]: https://issues.xnat.org/browse/CS-454
 
 ## 1.3.2